arm64: dts: r8a7795-salvator-xs: Add HSCIF1 device support
This patch adds the pin control for HSCIF1, and supports connection with
the Debug Serial-1 (CN26) for Salvator-X 2nd version board on R8A7795
ES2.0 SoC.

SCIF1 and HSCI1 are sharing the pin connected to the Debug Serial-1
(CN26) on Salvator-X board, and it is necessary to use it by exclusion.

As for the default of this DeviceTree, SCIF1 is connected.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>