rcar-vin: move functions regarding scaling
In preparation of refactoring the scaling code move the code regarding
scaling to to the top of the file to avoid the need to add forward
declarations. No code is changed in this commit only whole functions
moved inside the same file.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Reviewed-by: Hans Verkuil <hans.verkuil@cisco.com>
(cherry picked from commit 780c771426d82ff74faedf653090211fe5c042f5)
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>