iommu/ipmmu-vmsa: Reserve the 8th MMU context for MMNGR driver
This patch is to reserve the 8th MMU context for MMNGR driver which is
involved with Renesas Multimedia package.

MMNGR driver supports address space extension using IPMMU PMB
so that some media related IPs such as VCP4 make use of
40-bit physical address space.
However, MMNGR driver is no longer able to use IPMMU PMB because of
IPMMU H/W restriction No.71 #4 (see below for the details).

In order to support address space extension, MMNGR driver makes
use of the 8th one of 8 MMU contexts instead of PMB.
It is essential to share MMU resources b/w ipmmu-vmsa and MMNGR
driver from now on.

In case of CONFIG_RCAR_IPMMU_SHARE=y, the resource assignement is:
  - MMUn (n = 0 to 6) context is for ipmmu-vmsa
  - MMUn (n = 7) context is for MMNGR driver

In case of CONFIG_RCAR_IPMMU_SHARE=n, the assignement is:
  - MMUn (n = 0 to 7) context is for ipmmu-vmsa

The setting of this macro is 'n' by default and the location of
that macro is here.

  Device Drivers  --->
    IOMMU Hardware Support  --->
      Renesas VMSA-compatible IPMMU
        Share IPMMU resources with MMNGR driver [*]

----
IMPORTANT:

When the following conditions are met, a user need to set this macro
to 'y' by themselves.

  A) Use MMNGR driver with IPMMU MMU enabled
  B) Use IOMMU (CONFIG_IPMMU_VMSA=y)

<No.71 #4>
The address translation processing in the identical hierarchy
when uTLB which uses a PMB change exists, more than one,
when it forms, there is a possibility which causes a mis-conversion
and a stall.

Signed-off-by: Tomoharu Fukawa <tomoharu.fukawa.eb@renesas.com>
Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>