soc: renesas: rcar-sysc: Avoid the conflict of SYSC and CPU Hotplug
In very rare cases, if CPU Hotplug and SYSC control are processed
at same time, CPU stall will occur. It is caused by R-Car HW spec.
For avoid this issue, it adds exclusion control with mutex in sysc
driver.

This patch is a bad idea that may cause a deadlock in the future.
If SYSC control by pm_runtime is called from cpu notify, deadlock
will occur.(It doesn't occur now)  However, there are no other
workarounds, it needs to fix this now.

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>