v4l: vsp1: Fix completed check in write back sequence
There was an error in the writeback sequence below.
When the display completion notification was not issued,
the count value of writeback was decremented. As a result,
the wait processing of writeback was executed incorrectly
even though display was not completed. This patch corrects it.

In addition, the timing to read register reflected by display list
should be read at display start interrupt. Therefore, this patch
moves the procedure of writeback state transition from frame end
interrupt to display start interrupt.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>